% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recurrent_moe.r
\name{fit,RecurrentMoEModel,tf_dataset-method}
\alias{fit,RecurrentMoEModel,tf_dataset-method}
\title{fit}
\usage{
\S4method{fit}{RecurrentMoEModel,tf_dataset}(
  model,
  x,
  batch_size = 128L,
  epochs = 100L,
  learning_rate = 0.001,
  match_loss_type = "mse",
  compile = TRUE,
  training = TRUE
)
}
\arguments{
\item{model}{a VaeModel object, initialized by `new('VaeModel', model = VaeModel(...))`}

\item{x}{a tf_dataset object}

\item{batch_size}{Batch size (default: 256L)}

\item{epochs}{Number of training epochs (default: 500L)}

\item{learning_rate}{Learning rate (default: 1e-4)}

\item{compile}{Whether or not compile the tensorflow model (default: TRUE)}

\item{beta}{a weight for KL loss (default: 5e-5)}
}
\value{
a VaeModel
}
\description{
Fit a VaeModel
}
