% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsic_loss.r
\name{hsic_loss}
\alias{hsic_loss}
\title{Hilbert-Schmidt Independence Criteria (HSIC)}
\usage{
hsic_loss(x, y)
}
\arguments{
\item{x}{input data}

\item{y}{input data}
}
\value{
HSIC loss
}
\description{
Compute HSIC loss beteween two matrices
}
\references{
Tsai, Y.-H. H., Bai, S., Morency, L.-P. & Salakhutdinov, R.
A Note on Connecting Barlow Twins with Negative-Sample-Free Contrastive Learning. Arxiv (2021).

Weinberger, E. & Lee, S.-I. Transferable representations of single-cell transcriptomic data. Biorxiv 2021.04.13.439707 (2021) doi:10.1101/2021.04.13.439707.
}
