% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recurrent_moe.r
\name{prepare_data,RecurrentMoEModel,SingleCellExperimentList-method}
\alias{prepare_data,RecurrentMoEModel,SingleCellExperimentList-method}
\title{prepare_data}
\usage{
\S4method{prepare_data}{RecurrentMoEModel,SingleCellExperimentList}(
  model,
  x,
  batch_field = c("batch", "batch"),
  assay_field = c("X", "X"),
  weight = NULL,
  cell_type = NULL,
  ...
)
}
\description{
Prepare dataset for training a model
}
\author{
Wuming Gong (gongx030@umn.edu)
}
